no reason this needs to be part of class, move to conf
