+++ src/omni_epd/virtualepd.py
@@ -18,10 +18,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 """
 
-import sys
 import json
 import importlib
-import importlib.util
 import logging
 from PIL import Image, ImageEnhance
 from . conf import EPD_CONFIG, IMAGE_DISPLAY, IMAGE_ENHANCEMENTS
@@ -185,17 +183,6 @@ class VirtualEPD:
     # returns package.device name
     def getName(self):
         return self.__str__()
-
-    # helper method to check if a module is (or can be) installed
-    @staticmethod
-    def check_module_installed(moduleName):
-        result = False
-
-        # check if the module is already loaded, or can be loaded
-        if(moduleName in sys.modules or (importlib.util.find_spec(moduleName)) is not None):
-            result = True
-
-        return result
 
     # REQUIRED - a list of devices supported by this class, format is {pkgname.devicename}
     @staticmethod
